+++ orb/analytics/urls.py
@@ -12,4 +12,6 @@ urlpatterns = patterns('',
     url(r'^tag/(?P<id>\d+)/$', 'orb.analytics.views.tag_view', name="orb_analytics_tag"),
     url(r'^tag/(?P<id>\d+)/download/(?P<year>\d+)/(?P<month>\d+)/$', 'orb.analytics.views.tag_download', name="orb_analytics_download"),
     url(r'^resource/(?P<id>\d+)/$', 'orb.analytics.views.resource_view', name="orb_analytics_resource"),
+    
+    url(r'^review/$', 'orb.analytics.views.review_view', name="orb_analytics_review"),
 )